djangovue/requirements.txt
aioredis==1.3.1
argon2-cffi @ file:///C:/ci/argon2-cffi_1596828585465/work
asgiref==3.3.1
astroid==2.4.2
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.4.0
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
autobahn==21.3.1
Automat==20.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
bleach @ file:///tmp/build/80754af9/bleach_1600439572647/work
blinker==1.4
brotlipy==0.7.0
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1605538112425/work
channels==3.0.4
channels-redis==3.3.0
chardet @ file:///C:/ci/chardet_1605303225733/work
click==7.1.2
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
conda==4.9.2
conda-package-handling @ file:///C:/ci/conda-package-handling_1603003327818/work
constantly==15.1.0
cryptography==3.4.7
cycler==0.10.0
daphne==3.0.2
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.3
Django==3.1.4
django-allauth==0.45.0
django-ckeditor==6.0.0
django-cors-headers==3.11.0
django-crispy-forms==1.10.0
django-dajax==0.9.2
django-dajaxice==0.7
django-form-extensions==0.1.2b4
django-js-asset==1.2.2
django-markdownx==3.0.1
django-mptt==0.13.0
django-pagedown==2.2.0
django-resized==0.3.11
django-summernote==0.8.11.6
django-tinymce==3.2.0
django-wysiwyg==0.8.0
djangorestframework==3.12.4
dnspython==2.1.0
docutils==0.16
email-validator==1.1.2
entrypoints==0.3
fanstatic==1.1
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
hiredis==2.0.0
html5lib==1.1
hyperlink==21.0.0
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
incremental==21.3.0
iniconfig==1.1.1
ipykernel @ file:///C:/ci/ipykernel_1596190155316/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1604083276484/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1601490159889/work
isort==5.7.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
joblib==1.0.0
js.videojs==7.5.5
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1598884538475/work
jupyter-core @ file:///C:/ci/jupyter_core_1606149073702/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
kivymd==0.104.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lxml==4.6.2
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
msgpack==1.0.2
mysql-connector==2.2.9
mysql-connector-python==8.0.26
mysql-connector-python-rf==2.2.2
nbclient @ file:///tmp/build/80754af9/nbclient_1602783176460/work
nbconvert @ file:///C:/ci/nbconvert_1601914925608/work
nbformat @ file:///tmp/build/80754af9/nbformat_1602783287752/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1606153767164/work
notebook @ file:///C:/ci/notebook_1607976481874/work
numpy==1.19.4
oauthlib==3.1.1
packaging @ file:///tmp/build/80754af9/packaging_1607971725249/work
pandas==1.1.5
pandocfilters @ file:///C:/ci/pandocfilters_1605102497129/work
parso @ file:///tmp/build/80754af9/parso_1607623074025/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.0.1
platformdirs==2.3.0
pluggy==1.0.0
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1606344362066/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.17.3
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments @ file:///tmp/build/80754af9/pygments_1607368905949/work
PyJWT==2.1.0
pylint==2.6.0
pymongo==3.12.1
PyMySQL==1.0.2
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
pyparsing==2.4.7
pypiwin32==223
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pyrsistent @ file:///C:/ci/pyrsistent_1600141795814/work
PySocks @ file:///C:/ci/pysocks_1605287845585/work
pytest==6.2.5
pytest-asyncio==0.15.1
pytest-django==4.4.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
pywin32==227
pywinpty==0.5.7
pyzmq==20.0.0
qtconsole @ file:///tmp/build/80754af9/qtconsole_1600870028330/work
QtPy==1.9.0
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
ruamel-yaml==0.15.87
scikit-learn==0.24.0
scipy==1.5.4
selenium==3.141.0
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
service-identity==21.1.0
shutilwhich==1.1.0
sip==4.19.13
six @ file:///C:/ci/six_1605187374963/work
sklearn==0.0
soupsieve==2.1
SQLAlchemy==1.3.22
sqlparse==0.4.1
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
tornado @ file:///C:/ci/tornado_1606942392901/work
tqdm @ file:///tmp/build/80754af9/tqdm_1605303662894/work
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
Twisted==21.7.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
virtualenv==20.8.1
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
WebOb==1.8.7
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
wincertstore==0.2
wrapt==1.12.1
WTForms==2.3.3
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
zope.interface==5.4.0
